fastcore: Remove scaling parameter from LP10 (WIP)
The goal is to somehow remove the scaling parameter from LP10.
The scaling parameter is necessary to include the desired reactions
in the support set, but has the disadvantage that it is hard to
adjust in a way that ensures all the desired reactions are selected.
There is no simple way to detect whether all necessary reactions
were selected. This results in fastcore returning reaction subsets
that are not actually consistent according to fastcc because some
fluxes cannot go above epsilon.

There seems to be two solutions: 1) Make sure that all reactions
returned from find_sparse_mode() actually can go above epsilon.
This requires further refinement of the LP10 result, and it is not
certain how exactly; or 2) Allow fastcore() to find a subset that
is not consistent according to the epsilon threshold. When using
the model subsequently (FBA) there does not appear to be a need for
enforcing >epsilon reaction flux. find_sparse_mode() still has to
return a valid support in all cases, though, which is not the case
currently.